event board / register / withdrawal edit

+++ src/main/java/com/mp/cinepop/accountInsert/controller/AccountInsertController.java
@@ -82,27 +82,35 @@ public class AccountInsertController {
 	
 	 @PostMapping("mypage/withdrawal") 
 	 public String withdrawal_post(@RequestParam
-	 String pwd, @RequestParam String id, HttpSession session,HttpServletResponse response, Model model) throws NoSuchAlgorithmException, IOException {
+	 String pwd, @RequestParam String id,@RequestParam String pwd2, HttpSession session,HttpServletResponse response, Model model) throws NoSuchAlgorithmException, IOException {
 			/* id = (String)session.getAttribute("id"); */
 	 logger.info("hash 삭제, 파라미터 id={},pwd={}",id,pwd);
 	 
 	 int result=accountInsertService.loginCheck(id,pwd);
 	 if(result==accountInsertService.LOGIN_OK) {
-		 String salt=hash.makeNewSalt();
-		 String digest=hash.hashing(pwd, salt);
-		 int cnt =accountInsertService.deleteHash(id);
-		 int cnt2 =accountInsertService.withdrawAccount(id);
-		 if(cnt>0 && cnt2>0) {
-			 session.invalidate();
+		 if(pwd.equals(pwd2)) {
+             String salt=hash.makeNewSalt();
+             String digest=hash.hashing(pwd, salt);
+             int cnt =accountInsertService.deleteHash(id);
+             int cnt2 =accountInsertService.withdrawAccount(id);
+             if(cnt>0 && cnt2>0) {
+                session.invalidate();
+             }
+             return "mypage/withdrawalfin";
+		 }else {
+			 logger.info("비밀번호 확인 불일치");
 		 }
-	 }else if (result==accountInsertService.DISAGREE_PWD) {
+	 }else if (result==accountInsertService.DISAGREE_PWD || pwd.equals(pwd2)) {
 		 response.setContentType("text/html; charset=UTF-8");
 		 PrintWriter out = response.getWriter();
-		 out.print("<script>alert('비밀번호가 일치하지 않습니다.');history.go(-1);</script>");
+		 out.print("<script>");
+		 out.print("alert('비밀번호가 일치하지 않습니다.123');");
+		 out.print("history.back(-1);");
+		 out.print("</script>");
 		 out.flush();
 		 
 	 }
-	 return "mypage/withdrawalfin";
+	 return "mypage/withdrawal";
 	 }
